Fix remaining Kani compilation errors
- Fix type mismatches (u32 -> u64 for index, u64 -> u32 for lock_time)
- Fix field names (time -> timestamp, utxo_set -> new_utxo_set)
- Fix import paths (crate::transaction:: -> crate::types::)
- Fix function call (add missing witnesses parameter)
- Fix move errors (clone shared_prevout)
- Fix type comparison (u16 vs u32 for locktime_value)

+++ src/segwit.rs
@@ -772,7 +772,7 @@ mod kani_proofs {
             inputs.push(TransactionInput {
                 prevout: OutPoint {
                     hash: [0; 32],
-                    index: i as u32,
+                    index: i as u64,
                 },
                 script_sig: script,
                 sequence: 0xffffffff,
@@ -1147,8 +1147,7 @@ mod property_tests {
 #[cfg(kani)]
 mod kani_proofs_2 {
     use super::*;
-    use crate::block::Block;
-    use crate::transaction::Transaction;
+    use crate::types::{Block, Transaction};
     use kani::*;
 
     /// Kani proof: Witness commitment validation (Orange Paper Section 11.1)
